[C-4895] Fix profile screen verified badge
